Refactor out file operations into seperate helper Additional tests for MediaCompressor
